Prevent simulator from returning TPM_RC_RETRY
For certain TPM2 commands like ActivateCredential or PolicySecret, the
TPM is allowed to return TPM_RC_RETRY. In this simulator, this occurs
when dictionary attack prevention mechanisms are used. Specifically,
when checking to see if the maximum number of tries has been exceeded,
the simulator returns TPM_RC_RETRY to let certain values sync to nvdata.

While this can occur on real TPMs as well, it tends not to, and go-tpm
does not automatically handle this case. So we simply remove the above
feature.