proto: Add AttestationPolicy and MachineState
MachineState is an intermediate structure obtained from the event log.
MachineState represents certain properties on the machine, such as
platform firmware version and secure boot configuration. It allows
users to further parse go-attestation's output of an attest.Event slice
into a better summary of the state of the machine and to persist
the output of an event log in a protobuf for logging and future
analysis.

The policy proto represents a server implementation's configured
requirements for an attesting machine. It currently supports a
Secure Boot and platform policy.

The MachineState also allows server implementers to extend their policy
with a custom implementation using the raw_events field.