Skip tests when its Hash Algo is not supported
Some cases in TestReadPCRs, TestSign will be skipped if the hash algorithm is
not supported.
Also fixed TestReadPCRs. Previously it relys on the test PCR to be all zeros
at beginning of the test (true for the simulator, but not for a real TPM)
Remove the VerifyPSS with the specified salt length, as it behaves differently
between TPM implementations.