Change SealOpts to a struct instead of interface
SealOpts was previously an interface that implemented PCRsForSealing.
This did not allow users the option to seal to both a current PCR
selection and a target PCR selection without first reading PCRs manually
and merging into the target. This struct allows this use case more
easily and supports future options in the struct.