server: Support arbitrary hash algorithms for verification
Instead of hard-coding sha256, we now allow aribitary hash algorithms to
be use for verification. We will iteratate through a defined list of
supported hash algorithms, taking the first algorithm that produces a
valid MachineState.

We also now explictly check if the signing hash and the pcr hash are
supported by the library.

Signed-off-by: Joe Richey <joerichey@google.com>