Change CertifyOpts to an UnsealOpts struct
CertifyOpts was previously an interface that implemented CertifyPCRs.
This did not allow users the option to seal to both a current PCR
selection and an expected PCR selection without manually reading PCRs
and merging into the target. This struct allows this use case more
easily and supports future options in the struct.