Add SecureBootState support to ParseMachineState
This includes the Secure Boot variable (whether enabled), db, dbx, and
post-separator Authority. As with PlatformState, ParseMachineState may
return a zero value SecureBootState in MachineState if Secure Boot
parsing failed.

This commit also changes the output of
ParseMachineState to return *MachineStateError (implements error) so
that dependents can get a detailed view of which parts of the event log
fail parsing. As the type implements the Error interface, this is not
a breaking change.