Change MachineStateError to GroupedError
GroupedError allows the error that wraps a slice of errors to be reused
for Policy evaluation. This adds a helper function as well and fixes an
issue with the return value of ParseMachineState.

Previously, we always returned a non-nil MachineStateError from
ParseMachineState.  A zero value MachineStateError evaluate as
non-nil in error catching blocks, so we want to instead return
either nil or a pointer to a valid MachineStateError.