Add GRUB support, Bootloader option in VerifyOpts
GRUB measures each command it issues to PCR8. GRUB prefixes this with
context detailing whether it was a GRUB command, kernel commandline, or
module commandline.
This will allow users to decide whether to parse a
bootloader state if it is known ahead of time and supported (only GRUB for
now).
* Add EV_IPL event.
* Add Bootloader param to parsePCClientEventLog.
* Additionally, we add Bootloader as an option for VerifyOpts.