cel: Extend the PCR when adding an event to CEL
When calling AppendEvent, extend the target TPM's PCR for the event.
Also update the tests with a TPM handle and style and add helper method
for appending to a CEL.
Finally, change PCRs to use Debug and Application because test.GetTPM
can use a real TPM. This would clobber the values of 13 and 14 as the
test stands.