proto: Change RestartPolicy variants to use CamelCase
This is a backwards-incompatible change, which is fine because we
haven't released a version of go-tpm-tools containing `RestartPolicy`.
This change also means that the variants of RestartPolicy no longer
conform to the Protobuf style guide. An explanation for why we do this
is given in the comments in the `attest.proto` file.

Signed-off-by: Joe Richey <joerichey@google.com>