Add Cloud Logging; Update the service file
All launcher log will be sent to Cloud Logging as INFO
if running on GCE.
Stdout/Stderr of the container will still be redirected
to Stdout/Stderr and won't be sent to Cloud Logging.
Hosts now poweroff after the launcher exits.
Wrap the main function so the defered function can
be executed and return non-zero code.

Signed-off-by: Jiankun Lu <jiankun@google.com>