Add REST-based verifier.Client
This also adds a (very simple) networking test that I confirmed worked
on my local workstation. Note that this test will need to be changed in
the future, as the prod service will eventually stop having the fake
implementation.

Note that this test is also just skipped if there is no was to create a
default-authenticated client.

More tests should be added in a follow-up PR.

Signed-off-by: Joe Richey <joerichey@google.com>