Update ContainerRunner to use REST verifier
We update the LauncherSpec to store the projectID and region.

The ContainerRunner will now be able to use the REST-based
verifier.Client, based on the addr flag.
The flag will determine whether the endpoint is the
gRPC or REST verifier. The REST verifier is used if the flag is empty or
contains "http", otherwise the gRPC verifier is used.