Add an SEV-SNP attestation if possible
When getting a vTPM quote within an SEV-SNP VM, attestation verifiers
will also want to know if the SEV-SNP technology claim can be verified
with a hardware-rooted attestation.

We add "VerifyGceTechnology" for the GCENonHostInfo event that carries
which memory encryption technology the guest is using. SEV and SEV-ES do
not have attestation support in GCE, but SEV-SNP will.

The addition of the go-sev-guest dependency increases the minimum Go
version to 1.19.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>