launcher: Print kernel cmdline on builds
Previously, we would need to pull logs from the server to see how
exactly the cmdline gets passed through. Unfortunately, the cmdline
passed by GRUB, and as a result, measured in the TCG event log, has
surrounding quotes stripped out. So, we mutate the existing command line
to what GRUB passes to the kernel.