Add/update tests for images/launcher
Add a basic/smoke test which test some basic functionality of an
image.
Update some scripts so they can take in args like vm name and zone.
Add a test to check user-data won't be executed in the hardened
image.
Update cloudbuild steps to have ids so they'll be easier to read.

Signed-off-by: Jiankun Lu <jiankun@google.com>