Increase default systemd wait timeout to 900s
The systemd default timeout is 90s. This is often insufficient for
standing up the protected stateful partition, which relies on writing
dm-integrity tags for integrity protection. This can fail due to PD
throttling on smaller disks and a machine that reboots multiple times.
Additionally, this makes the debug image more usable on non-Confidential
VM machines (which often have fewer/timeshared vCPUs).

We benchmarked a lower bound for the timeout on GCE, around 900s.