[launcher] Fix token directory permission
Current token directory has permission "drw-r--r--", so only root
can access the token file inside. This is fine if the container
process is run as root. But if the container process is run as
a non-root user, it won't be able to  access the token file.
Change the permission so any process can access the token file.