Remove VerifyAttestation dependency on SNP/TDX
Previously, VerifyAttestation required passing a SNP or TDX attestation
when it detected the appropriate NonHostInfo event in the UEFI event
log. This introduces a weak verification requirement that gives no
guarantees in the resulting claim that the vTPM and SNP/TDX devices are
actually on the same machine.

Fixes #500 and #504.