Remove VerifyAttestation dependency on SNP/TDX (#523)
Previously, VerifyAttestation required passing a SNP or TDX attestation
when it detected the appropriate NonHostInfo event in the UEFI event
log. This introduces a weak verification requirement that gives no
guarantees in the resulting claim that the vTPM and SNP/TDX devices are
actually on the same machine.

Fixes #500. We still need to remove the client auto collection of SNP/TDX
attestation (#504), but that is a bit more work since we also need to
update the gotpm CLI to include SNP and TDX attestation collection.