implement EK-based Key attestation for the SVSM e-vTPM
The client server files for import_certify.go corresponds to the spec
for EK-based Key attestation with firmware version.
See https://trustedcomputinggroup.org/wp-content/uploads/EK-Based-Key-Attestation-with-TPM-Firmware-Version-V1-RC1_9July2025.pdf#page=8

This will be used for SVSM attestation in the next patches. This is
required for SVSM attestation because Coconut SVSM currently only
attests to a single EK.