Update end-to-end testing to use fake verifier (#607)
Updates the testworkloads to use the fake verifier, rather than the remote attestation service.  

A new launch spec variable, test-fake-verifier, is added and set to true in all test VMs. This sets the attestation client to an instance of the fake verifier.

The fake verifier has been updated to use hardcoded local keys, rather than pulling keys from the remote service, and to return test values for some fields.  Tests have been updated to test for these specific values.

This is part of a change to remove remote service dependency in the client end-to-end tests.